utils/requirements.txt
#requirements.txt
sentence-transformers==0.4.1.2
faiss-cpu
psutil
beautifulsoup4
matplotlib
